+++ src/middleware/req-body.js
@@ -1,7 +1,7 @@
 const debug = require("debug")("Blocks:BodyMiddleware")
-const { isEmpty } = require("@asd14/m")
 
-const InputValidationError = require("../errors/input.error")
+import { isEmpty } from "@asd14/m"
+import { InputValidationError } from "../errors/input"
 
 /**
  * { function_description }